Create subpackage for asyncio
